Replace RTMDet-R with YOLOX-OBB on the same runtime
Pure-PyTorch YOLOX-OBB for oriented banknote detection (CPU, CUDA, MPS),
keeping the data pipeline, augmentations, training loop, evaluation,
inference and ONNX tooling of the RTMDet-R runtime.

- Networks (--model): YOLOX-OBB-s of DDGRCF/YOLOX_OBB (DOTA weights,
  same state-dict layout as the original) and the official Megvii
  YOLOX nano / tiny / s with a 5-channel regression head (COCO weights,
  pretrained xywh filters kept, angle channel starts at zero).
- Loss and assignment of YOLOX-OBB: BCE objectness, BCE class with IoU
  target, 5 x rotated IoU, L1 in the last epochs; rotated SimOTA.
- YOLOX optimisation: SGD nesterov, yoloxwarmcos, ModelEMA 0.9998; the
  stage-2 (light augmentation) epochs are the YOLOX no-aug phase.
- ONNX export decodes inside the graph (boxes, scores) and writes a
  metadata contract; onnx_example.py is a torch-free reference client.
- HEIC images are read by the dataset (the external set has 59 of them).
- Fixes: f1 as numpy float broke log.jsonl, experiment-relative
  checkpoint paths in evaluate/infer.

+++ src/yolox_mps/yolox_obb/evaluation.py
@@ -102,7 +102,7 @@ def summarize(det_results, annotations, num_classes, thresholds: Sequence[float]
     r = eval_rbbox_map(operational, annotations, num_classes, 0.5, mode)
     precision = float(np.mean([r[f'precision{c}'] for c in range(num_classes)]))
     recall = float(np.mean([r[f'recall{c}'] for c in range(num_classes)]))
-    f1 = 2 * precision * recall / max(precision + recall, np.finfo(np.float32).eps)
+    f1 = float(2 * precision * recall / max(precision + recall, float(np.finfo(np.float32).eps)))
     res['precision@0.50'] = precision
     res['recall@0.50'] = recall
     res['f1@0.50'] = f1
